Crate radiant_rs

Modules

blendmodes
A set of predefined blendmodes for use with Layer::set_blendmode().

scene
Optional ultility wrapper to simplify scene handling.

utils
Various optional utility features including a Scene and mainloop helpers.

Structs

BlendMode
A blendmode for use with Layer::set_blendmode(). See blendmodes for a list of predefined modes.

Color
A color value consisting of four floating point values for the color channels red, green, blue and alpha.

Display
A target to render to, e.g. a window or full screen.

DisplayInfo
A struct describing a Display to be created.

Font
A font used for writing on a Layer.

FontInfo
A struct used to filter the result of Font::query_specific() or to describe a Font to be created from a system font via Font::from_info().

Input
Basic keyboard and mouse support.

Layer
A non-blocking, thread-safe drawing target.

Mat4
A 4x4 matrix.

Monitor
An individual monitor, returned from Display::monitors().

RenderContext
A thread-safe render-context.

Renderer
A renderer is used to render Layers or Scenes to the Display.

Sprite
A sprite used for drawing on a Layer.

Vec2
A 2-dimensional vector.

Vec3
A 3-dimensional vector.

Enums

ButtonState
The current state of a key or mousebutton.

Traits

VecType