Enum radiant_rs::Uniform

pub enum Uniform {
    Bool(bool),
    SignedInt(i32),
    UnsignedInt(u32),
    Float(f32),
    Mat4([[f32; 4]; 4]),
    Vec2([f32; 2]),
    Vec3([f32; 3]),
    Vec4([f32; 4]),
    Double(f64),
    DoubleMat4([[f64; 4]; 4]),
    DoubleVec2([f64; 2]),
    DoubleVec3([f64; 3]),
    DoubleVec4([f64; 4]),
    Texture(Texture),
}

A uniform value.

Uniforms are values that can be passed to Programs. Various types also implement the AsUniform trait and can be directly used with Program::set_uniform().

Variants
Bool(bool)SignedInt(i32)UnsignedInt(u32)Float(f32)Mat4([[f32; 4]; 4])Vec2([f32; 2])Vec3([f32; 3])Vec4([f32; 4])Double(f64)DoubleMat4([[f64; 4]; 4])DoubleVec2([f64; 2])DoubleVec3([f64; 3])DoubleVec4([f64; 4])Texture(Texture)