Struct radiant_rs::math::Mat4

pub struct Mat4<T: Debug + Float + NumCast = f32>(pub [[T; 4]; 4]);

A 4x4 matrix.

Methods

impl<T> Mat4<T> where T: Debug + Float + NumCast

fn new() -> Mat4<T>

Creates a zero matrix.

fn identity() -> Mat4<T>

Creates an identity matrix.

fn viewport(width: T, height: T) -> Mat4<T>

Creates viewport matrix mapping viewport top left to (0.0, 0.0) and bottom right to (width, height)

fn look_at<E, C, U>(eye: E, target: C, up: U) -> Mat4<T> where E: VecType<T>, C: VecType<T>, U: VecType<T>

Creates a look-at matrix with the given eye position, target position, and up-vector.

fn ortho<R>(rectangle: R, near: T, far: T) -> Mat4<T> where Rect<T>: From<R>

Creates an orthogonal projection matrix with the given rectangular bounds at the near and far clipping plane.

fn frustum<R>(rectangle: R, near: T, far: T) -> Mat4<T> where Rect<T>: From<R>

Creates a frustum projection matrix with the given rectangular bounds at the near clipping plane and rectangle * (far / near) at the far clipping plane.

fn perspective(fov_y: T, aspect: T, near: T, far: T) -> Mat4<T>

Creates a perspective projection matrix with the given field of view, aspect and near/far clipping planes.

fn translate<V>(&mut self, translation_vector: V) -> &mut Self where V: VecType<T>

Translate matrix by given vector.

fn scale<V>(&mut self, scaling_vector: V) -> &mut Self where V: VecType<T>

Scale matrix by given vector.

fn scale_at<P, V>(&mut self, position: P, scaling_vector: V) -> &mut Self where P: VecType<T>, V: VecType<T>

Scales at given position.

fn rotate(&mut self, radians: T) -> &mut Self

Rotates the origin around z.

fn rotate_at<P>(&mut self, position: P, radians: T) -> &mut Self where P: VecType<T>

Rotates around z at given position.

fn rotate_axis<V>(&mut self, radians: T, axis: V) -> &mut Self where Vec3<T>: From<V>

Rotates around axis.

fn rotate_axis_at<P, V>(&mut self,
position: P,
radians: T,
axis: V)
-> &mut Self where P: VecType<T>, Vec3<T>: From<V>

Rotates around axis at given position.

fn set(&mut self, other: Mat4<T>)

Sets the matrix value from another matrix.

fn get_rotation(&self) -> Mat4<T>

Returns a pure rotation matrix for given matrix

fn get_translation(&self) -> Vec3<T>

Returns the matrix's translation vector.

fn get_scale(&self) -> Vec3<T>

Returns the matrix's scaling vector.

fn get_euler(&self) -> Vec3<T>

Get rotation matrix euler angles