Struct radiant_rs::postprocessors::Basic

pub struct Basic { /* fields omitted */ }

A basic postprocessor that applies a Program to the given input once.

Postprocessors are used with Renderer::postprocess(). The associated type for this postprocessor is BlendMode and is expected as second argument to the postprocess() method

// Load a shader progam.
let my_program = Program::from_string(&rendercontext, include_str!("my_shader.fs")).unwrap();

// Create the postprocessor with the program.
let my_postprocessor = postprocessors::Basic::new(&rendercontext, my_program);

// ... in your renderloop...
renderer.postprocess(&my_postprocessor, &blendmodes::ALPHA, || {
    renderer.clear(Color::black());
    renderer.draw_layer(&my_layer, 0);
});

Methods

impl Basic

fn new(context: &RenderContext, program: Program) -> Self

Creates a new instance. The shader can use sheet*() to access the input texture.