Enum radiant_rs::scene::Op

pub enum Op {
    None,
    SetColor(LayerId, Color),
    SetBlendmode(LayerId, BlendMode),
    Draw(LayerId),
    Clear(LayerId),
    SetViewMatrix(LayerId, Mat4<f32>),
    TranslateViewMatrix(LayerId, f32, Vec2<f32>),
    RotateViewMatrix(LayerId, f32, f32),
    RotateViewMatrixAt(LayerId, f32, Vec2<f32>, f32),
    ScaleViewMatrix(LayerId, f32, Vec2<f32>),
    SetModelMatrix(LayerId, Mat4<f32>),
    TranslateModelMatrix(LayerId, f32, Vec2<f32>),
    RotateModelMatrix(LayerId, f32, f32),
    RotateModelMatrixAt(LayerId, f32, Vec2<f32>, f32),
    ScaleModelMatrix(LayerId, f32, Vec2<f32>),
}

A scene-operation.

Each time a scene is drawn, all of the registered scene operations will be performed. Transformation-, rotation- and scaling-operations take a delta-factor which determines how the operation is affected by frame delta. A value of 1.0 will scale the relevant parts of the operation by frame-delta, a value of 0.0 will not apply any modifications to the operation.

Variants
None

No op

SetColor(LayerId, Color)

Set multiplication color for given layer.

SetBlendmode(LayerId, BlendMode)

Sets the blendmode for given layer.

Draw(LayerId)

Draws given layer.

Clear(LayerId)

Clears given layer.

SetViewMatrix(LayerId, Mat4<f32>)

Set a layer's view matrix.

TranslateViewMatrix(LayerId, f32, Vec2<f32>)

Translate view matrix by given vector.

RotateViewMatrix(LayerId, f32, f32)

Rotate view matrix by given radians.

RotateViewMatrixAt(LayerId, f32, Vec2<f32>, f32)

Rotate view matrix by given radians around given origin.

ScaleViewMatrix(LayerId, f32, Vec2<f32>)

Scale view matrix by given vector.

SetModelMatrix(LayerId, Mat4<f32>)

Set a layer's model matrix.

TranslateModelMatrix(LayerId, f32, Vec2<f32>)

Translate model matrix by given vector.

RotateModelMatrix(LayerId, f32, f32)

Rotate model matrix by given radians.

RotateModelMatrixAt(LayerId, f32, Vec2<f32>, f32)

Rotate view matrix by given radians around given origin.

ScaleModelMatrix(LayerId, f32, Vec2<f32>)

Scale model matrix by given vector.