Struct radiant_rs::Angle

pub struct Angle<T: Debug + Float + NumCast = f32>(pub T);

An Angle between -PI and PI

Methods

impl<T> Angle<T> where T: Debug + Float + NumCast

fn to_radians(self: &Self) -> T

Returns the angle's value in radians.

fn to_degrees(self: &Self) -> T

Returns the angle's value in degrees.

fn from_radians(radians: T) -> Angle<T>

Creates an angle from a radians value.

fn from_degrees(degrees: T) -> Angle<T>

Creates an angle from a degrees value.

fn normalize(self: &mut Self) -> &mut Self

Mutates self to its normalized representation

fn align_with(self: &mut Self, target: &Angle<T>) -> &mut Self

Mutates self so that subtracting the target will yield the smallest directional angle between self and target