Struct radiant_rs::Display

pub struct Display { /* fields omitted */ }

A target to render to, e.g. a window or full screen.

Methods

impl Display

fn builder() -> DisplayBuilder

Returns a display builder for display construction.

let display = Display::builder().dimensions((640, 480)).vsync().title("Window!").build();

fn set_title(self: &Self, title: &str)

Sets the window title.

fn show(self: &Self)

Makes the previously hidden window visible.

fn hide(self: &Self)

Hides the window.

fn prepare_frame(self: &Self)

Prepares a frame for rendering.

fn clear_frame(self: &Self, color: Color)

Prepares a frame for rendering and clears it.

fn swap_frame(self: &Self)

Swaps current drawing frame with visible frame.

fn grab_cursor(self: &Self)

Enables cursor grab mode. While in this mode, the mouse cursor will be hidden and constrained to the window. Additionally, Input will be able to provide mouse movement deltas and allow mouse coordinates to exceed the window-bounds.

Grab mode will be temporarily released when the window loses focus and automatically restored once it regains focus.

fn hide_cursor(self: &Self)

Hides the mouse cursor while it is inside the window.

fn free_cursor(self: &Self)

Releases a previously grabbed or hidden cursor and makes it visible again.

fn dimensions(self: &Self) -> Point2<u32>

Returns the window dimensions.

fn monitor(index: u32) -> Option<Monitor>

Returns monitor details for given monitor id.

fn monitors() -> Vec<Monitor>

Returns a vector of available monitors.

fn from_glium(display: Display) -> Display

Takes a glium::Display and returns a radiant::Display.

fn poll_events(self: &Self) -> &Self

Polls for events like keyboard or mouse input and changes to the window. See Input for basic keyboard and mouse support.

fn was_closed(self: &Self) -> bool

Returns true once after the attached window was closed