Struct radiant_rs::Input

pub struct Input { /* fields omitted */ }

Basic keyboard and mouse support.

Methods

impl Input

fn new(display: &Display) -> Self

Creates a new instance.

fn iter(self: &Self) -> InputIterator

Returns an iterator over all keys and buttons.

fn mouse(self: &Self) -> (i32, i32)

Returns current mouse coordinates relative to the window.

fn mouse_delta(self: &Self) -> (i32, i32)

Returns mouse delta coordinates since last Display::poll_events().

fn down(self: &Self, key: InputId) -> bool

Returns true if given key is down/pressed.

fn pressed(self: &Self, key: InputId, report_repeats: bool) -> bool

Returns true if given key was just pressed or repeated due to still being held down (if report_repeats is true).

fn up(self: &Self, key: InputId) -> bool

Returns true if given key is up/released.

fn released(self: &Self, key: InputId) -> bool

Returns true if given key was just released.

fn state(self: &Self, key: InputId) -> InputState

Returns InputState for given key.