Struct radiant_rs::Scene

pub struct Scene<'a> {
    // some fields omitted
}

[WIP] Poke with long stick. Wear gloves.

Methods

impl<'a> Scene<'a>

fn new(max_sprites: u32, dimensions: (u32, u32)) -> Scene<'a>

create a new scene instance

fn op(&self, op: Operation) -> OperationId

push a layer operation on the scene operation stack

fn ops(&self, ops: &[Operation]) -> &Self

fn clear(&self) -> &Self

clear operation stack

fn sprite_transformed(&self, layer_id: LayerId, sprite: Sprite, frame_id: u32, x: f32, y: f32, color: Color, rotation: f32, scale_x: f32, scale_y: f32)

draws a sprite with given rotation and scaling onto given layer

fn sprite(&self, layer_id: LayerId, sprite: Sprite, frame_id: u32, x: f32, y: f32, color: Color)

draws a sprite onto given layer

fn add_layer(&self) -> LayerId

create and add a layer to the scene