Struct radiant_rs::Texture

pub struct Texture { /* fields omitted */ }

A texture to draw or draw to.

Textures serve as drawing targets for userdefined Postprocessors or custom Programs. A texture can also be drawn with Renderer::rect().

Methods

impl Texture

fn builder(context: &RenderContext) -> TextureBuilder

Returns a textu builder for texture construction.

let tex = Texture::builder(&rendercontext)
                    .dimensions((640, 480))
                    .magnify(TextureFilter::Nearest)
                    .minify(TextureFilter::Linear)
                    .build()
                    .unwrap();

fn new(context: &RenderContext, width: u32, height: u32) -> Self

Creates a new texture with given dimensions. The texture will use linear interpolation for magnification or minification and internally use the F16F16F16F16 format.

fn filtered(context: &RenderContext,
width: u32,
height: u32,
minify: TextureFilter,
magnify: TextureFilter)
-> Self

Creates a new texture with given dimensions and filters. It will internally use the F16F16F16F16 format.

fn clone_with_options(self: &Self,
minify: TextureFilter,
magnify: TextureFilter,
wrap: TextureWrap)
-> Self

Clones texture with new filters and wrapping function. Both source and clone reference the same texture data.

fn clear(self: &Self, color: Color)

Clears the texture with given color.

fn dimensions(self: &Self) -> Point2<u32>

Returns the dimensions of the texture.