Struct radiant_rs::Vec2

pub struct Vec2<T: Debug + Float = f32>(pub T, pub T);

A 2-dimensional vector.

Methods

impl<T> Vec2<T> where T: Debug + Float

fn new() -> Self

Creates a new instances.

fn len(self: &Self) -> T

Returns the length of the vector

fn dot(self: &Self, other: &Self) -> T

Returns the dot-product of the vectors.

fn to_radians(self: &Self) -> T

Returns the direction of the vector in radians.

fn to_degrees(self: &Self) -> T

Returns the direction of the vector in degrees.

fn to_angle(self: &Self) -> Angle<T>

Returns the direction of the vector as an angle instance.

fn from_radians(radians: T) -> Self

Creates a unit-vector from the angle given in radians.

fn from_degrees(degrees: T) -> Self

Creates a unit-vector from the angle given in degrees.

fn from_angle(angle: Angle<T>) -> Self

Creates a unit-vector from given angle.

fn normalize(self: Self) -> Self

Normalizes the vector.

fn left(self: Self) -> Self

The left pointing normal of the vector.

fn right(self: Self) -> Self

The right pointing normal of the vector.

fn extend(self: Self, extension_len: T) -> Self

Extends the vector by given length.

fn outbound(self: &Self, bounding: Rect<T>) -> Option<Self>

Returns outbound vector for this point and given bounding box. Subtracting it from this point will result in a point on the bounding box.

fn is_zero(self: &Self) -> bool

Returns true if the vecor is a zero-vector.

fn distance(self: &Self, other: &Self) -> T

Returns distance to other point.