Struct radiant_rs::support::FontBuilder

#[must_use]
#[must_use]

pub struct FontBuilder<'a> { /* fields omitted */ }

A font builder.

Obtained from Font::builder().

let my_font = Font::builder(&rendercontext).family("Arial").size(16.0).build().unwrap();

Methods

impl<'a> FontBuilder<'a>

fn family(self: Self, family: &str) -> Self

Sets a family for the Font. The font will be retrieved from the operating system. Mutually exclusive with file().

fn file(self: Self, file: &'a str) -> Self

Sets file for the Font to be loaded from. Mutually exclusive with family().

fn italic(self: Self) -> Self

Flags the Font to be italic.

fn oblique(self: Self) -> Self

Flags the Font to be oblique.

fn monospace(self: Self) -> Self

Flags the Font to be monospace.

fn size(self: Self, size: f32) -> Self

Sets the fontsize.

fn build(self: Self) -> Result<Font>

Returns the constructed Font instance.