Trait radiant_rs::Postprocessor

pub trait Postprocessor {
    type T;
    fn target(&self) -> &Texture;
    fn draw(&self, renderer: &Renderer, args: &Self::T);

    fn process(&self, renderer: &Renderer, args: &Self::T) { ... }
}

A custom postprocessor.

Postprocessing happens in three steps:

first, target() is invoked and expected to return the input texture target (from where the postprocessor intends to read input data).

process() is invoked and expected to perform the neccessary processing excluding the final draw operation.

draw() is invoked. At this point the renderer has already restored the drawing target so that this method is only required to draw the postprocessing result.

While you could combine processing and drawing within draw(), it is recommended to separate these as future versions might expand on this functionality.

Associated Types

type T

Custom type for the args parameter supplied to process() and draw().

Required Methods

fn target(&self) -> &Texture

Expected to return a texture to draw to.

fn draw(&self, renderer: &Renderer, args: &Self::T)

Expected to draw final result. Draws issued within this function will target the rendertarget that the postprocessor is supposed to render to.

Provided Methods

fn process(&self, renderer: &Renderer, args: &Self::T)

Expected to processes input data. Draws issued within this function will target the texure returned by target() unless overridden via Renderer::render_to(). This function is provided as you could also do this in draw().