Struct radiant_rs::utils::AVec

pub struct AVec<T> {
    // some fields omitted
}

Vector supporting multiple non-overlapping writers or multiple readers. Blocks during resizes only. !todo panics on concurrent reads+writes

Methods

impl<T> AVec<T>

fn new(initial_capacity: usize) -> AVec<T>

Creates a new instance with given inital capacity

fn push(self: &Self, value: T) -> usize

Adds an element to the vector and returns the insert position.

fn map<'a>(&'a self, size: usize) -> AVecMapGuard<'a, T>

Maps a slice of the vector for write access. Faster than individual pushes for size > 2.

fn len(self: &Self) -> usize

Returns the current length of the vector.

fn grow(self: &Self, required_capacity: usize)

Grow vector by given number of elements.

fn clear(self: &Self)

Clears the vector.

fn get<'a>(&'a self) -> AVecReadGuard<'a, T>

Maps the entire vector for read access.

fn capacity(self: &Self) -> usize

Returns current capacity. The vector will reallocate once this is exceeded.