Struct radiant_rs::scene::Scene

pub struct Scene { /* fields omitted */ }

A container for layers, sprites and fonts intended to simplify usage.

Methods

impl Scene

fn new(context: &RenderContext) -> Self

Create a new scene instance.

fn op(&self, op: Op) -> OpId

Push a layer operation on the scene operation stack.

fn ops(&self, ops: &[Op]) -> &Self

Push multiple operations on the scene operation stack.

fn clear(&self) -> &Self

Clear operation stack.

fn sprite(&self,
layer_id: LayerId,
sprite_id: SpriteId,
frame_id: u32,
x: f32,
y: f32,
color: Color)
-> &Self

Draws a sprite onto given layer.

fn sprite_transformed(&self,
layer_id: LayerId,
sprite_id: SpriteId,
frame_id: u32,
x: f32,
y: f32,
color: Color,
rotation: f32,
scale_x: f32,
scale_y: f32)
-> &Self

Draws a sprite with given rotation and scaling onto given layer.

fn write(&self,
layer_id: LayerId,
font_id: FontId,
text: &str,
x: f32,
y: f32)
-> &Self

Writes a string onto given layer.

fn register_layer(&self, width: u32, height: u32) -> LayerId

Create and register a layer to the scene.

fn register_sprite_from_file(self: &Self, file: &str) -> SpriteId

Create and register a sprite to the scene

fn register_sprite(self: &Self, sprite: Sprite) -> SpriteId

Register a sprite for the scene.

fn register_font(self: &Self, font: Font) -> FontId

Register a font for the scene.