Struct radiant_rs::Color

pub struct Color(pub f32, pub f32, pub f32, pub f32);

A color value consisting of four floating point values for the color channels red, green, blue and alpha.

Various drawing methods accept color instances to be used as multiplicators in the drawing process, e.g. Sprite::draw() allows multiplying the sprite- texture's color channels by given color.

Methods

impl Color

fn transparent() -> Color

Creates a new instance with a channels set to zero.

fn alpha(alpha: f32) -> Color

Creates a new instance with color channels set to one and the alpha channel set to given value.

fn alpha_mask(alpha: f32) -> Color

Creates a new instance with color channels set to zero and the alpha channel set to given value.

fn alpha_pm(alpha: f32) -> Color

Creates a new instance with all channels set to given value.

fn lightness(value: f32) -> Color

Creates a new instance with color channels set to given value and the alpha channel set to one.

fn from_hsl(hue: f32, saturation: f32, lightness: f32, alpha: f32) -> Color

Creates a new instance from given HSL (range 0.0 - 1.0)

fn from_temperature(temperature: f32, alpha: f32) -> Color

Creates a new instance from given color-temperature (~1000 to ~40000).

Based on http://www.tannerhelland.com/4435/convert-temperature-rgb-algorithm-code/

fn r(self: &Self) -> f32

Returns value of the instance's red channel.

fn g(self: &Self) -> f32

Returns value of the instance's green channel.

fn b(self: &Self) -> f32

Returns value of the instance's blue channel.

fn a(self: &Self) -> f32

Returns value of the instance's alpha channel.

fn set(self: &mut Self, value: Color)

Sets the instance's channels from another color object.

fn set_r(self: &mut Self, value: f32) -> Color

Sets a value for the instance's red channel

fn set_g(self: &mut Self, value: f32) -> Color

Sets a value for the instance's green channel.

fn set_b(self: &mut Self, value: f32) -> Color

Sets a value for the instance's blue channel.

fn set_a(self: &mut Self, value: f32) -> Color

Sets a value for the instance's alpha channel.

fn scale(self: &mut Self, scaling: f32) -> Color

Multiplies the instance's color channels by given scaling factor. Does not modify alpha.

fn as_tuple(self: &Self) -> (f32, f32, f32, f32)

Returns the instance's channels as a tuple.

fn white() -> Color

fn black() -> Color

fn red() -> Color

fn green() -> Color

fn blue() -> Color

fn yellow() -> Color

fn cyan() -> Color

fn purple() -> Color