Struct radiant_rs::Font

pub struct Font { /* fields omitted */ }

A font used for writing on a Layer.

Fonts can be created from files, registered system fonts or existing font objects. When creating fonts from system fonts, a FontInfo structure can be used to define requirements for the font, e.g. "any available monospace font".

In addition to the usual properties of a font, radiant also assigns a fixed color and size to each font object. Instead of modifying these properties, you can clone a new font with modified values using Font::with_color() and/or Font::with_size().

Methods

impl Font

fn from_file(context: &RenderContext, file: &str) -> Font

Creates a font instance from a file

fn from_info(context: &RenderContext, info: FontInfo) -> Font

Creates a new font instance from given FontInfo struct

fn query_all() -> Vec<String>

Returns the names of all available system fonts

fn query_specific(info: FontInfo) -> Vec<String>

Returns the names of all available system fonts with the given properties (e.g. monospace)

fn with_size(self: &Self, size: f32) -> Font

Returns a new font instance with given size

fn with_color(self: &Self, color: Color) -> Font

Returns a new font instance with given color

fn write(self: &Self, layer: &Layer, text: &str, position: Point2) -> &Font

Write to given layer

fn write_wrapped(self: &Self,
layer: &Layer,
text: &str,
position: Point2,
max_width: f32)
-> &Font

Write to given layer. Breaks lines after max_width pixels.

fn write_transformed(self: &Self,
layer: &Layer,
text: &str,
position: Point2,
max_width: f32,
rotation: f32,
scale: Vec2)
-> &Font

Write to given layer. Breaks lines after max_width pixels and applies given rotation and scaling.