Struct radiant_rs::Renderer

pub struct Renderer { /* fields omitted */ }

A renderer is used to render Layers or Scenes to the Display.

The renderer itself is not thread-safe. Instead, draw or write onto layers (from any one or more threads) and present those layers using the renderer once your threads have concluded drawing.

Alternatively to directly drawing on layers, Scene provides a higher level abstraction.

Methods

impl<'a> Renderer

fn new(display: &Display) -> Self

Returns a new renderer instance.

fn context(self: &Self) -> RenderContext

Returns a reference to the renderers' context. The RenderContext is thread-safe and required by Font and Sprite to create new instances.

fn prepare_target(self: &Self)

Prepares the target for drawing without clearing it.

fn clear_target(self: &Self, color: Color)

Prepares the target and clears it with given color.

fn swap_target(self: &Self)

Finishes drawing and swaps front with backbuffer.

fn draw_scene(self: &Self, scene: &Scene, per_frame_multiplier: f32) -> &Self

Draws given scene.

fn draw_layer(self: &Self, layer: &Layer) -> &Self

Draws given layer.