Struct radiant_rs::Renderer

pub struct Renderer { /* fields omitted */ }

A renderer is used to render Layers or Scenes to the Display.

The renderer itself is not thread-safe. Instead, draw or write onto layers (from any one or more threads) and present those layers using the renderer once your threads have concluded drawing.

Alternatively to directly drawing on layers, Scene provides a higher level abstraction.

Methods

impl<'a> Renderer

fn new(display: &Display) -> Result<Self, Error>

Returns a new renderer instance.

fn context(self: &Self) -> RenderContext

Returns a reference to the renderers' context. The RenderContext is thread-safe and required by Font and Sprite to create new instances.

fn set_target<T>(self: &Self, target: &T) where T: RenderTarget

Sets a new rendering target. Valid targets are the display and textures.

fn clear(self: &Self, color: &Color)

Clears the current target.

fn draw_scene(self: &Self, scene: &Scene, per_frame_multiplier: f32) -> &Self

Draws given scene to the current target..

fn draw_layer(self: &Self, layer: &Layer) -> &Self

Draws given layer to the current target..

fn draw_layer_processed<T>(self: &Self,
layer: &Layer,
postprocessor: &mut T)
-> &Self where T: Postprocessor

Draws given layer to the current target using given postprocessor.

fn draw_rect<T, S>(self: &Self,
position: T,
dimensions: S,
blendmode: BlendMode,
program: Option<&Program>,
texture: Option<&Texture>)
-> &Self where Vec2<f32>: From<T> + From<S>

Draws a rectangle to the current target. The optionally specified texture is available via sheet*() in the shader. Note that you can pass custom textures via the optional shader program's uniforms.

fn default_program(self: &Self) -> &Program

Returns a reference to the default rendering program.