Struct radiant_rs::Renderer

pub struct Renderer<'a> {
    // some fields omitted
}

A renderer is used to render Layers or Scenes to the Display.

The renderer itself is not thread-safe. Instead, draw or write onto layers (from any one or more threads) and present those layers using the renderer once your threads have concluded drawing.

Alternatively to directly drawing on layers, Scene provides a higher level abstraction.

Methods

impl<'a> Renderer<'a>

fn new(display: &Display, max_sprites: u32) -> Self

Returns a new renderer instance. Currently, the maximum number of expected sprites per layer has to be set during instance creation. Future versions will lift this requirement.

fn context(&self) -> Arc<RenderContext<'a>>

Returns a reference to the renderers' context. The RenderContext is thread-safe and required by Font and Sprite to create new instances.

fn prepare_target(&self)

Prepares a new target for drawing without clearing it.

fn clear_target(&self, color: Color)

Prepares a new target and clears it with given color.

fn swap_target(&self)

Finishes drawing and swaps the drawing target to front.

fn draw_scene(&self, scene: &Scene) -> &Self

Draws given scene.

fn draw_layer(&self, layer: &Layer) -> &Self

Draws given layer.