Struct radiant_rs::Sprite

pub struct Sprite {
    pub anchor: Point2,
    // some fields omitted
}

A sprite used for drawing on a Layer.

Sprites are created from spritesheets containing one or more frames. To determine frame dimensions, Sprite::from_file() expects sprite sheet file names to follow a specific pattern. (Future versions will add more configurable means to load sprites.)

Fields
anchor: Point2

Defines the sprite origin. Defaults to (0.5, 0.5), meaning that the center of the sprite would be drawn at the coordinates given to Sprite::draw(). Likewise, (0.0, 0.0) would mean that the sprite's top left corner would be drawn at the given coordinates.

Methods

impl<'a> Sprite

fn from_file(context: &RenderContext, file: &str) -> Sprite

Creates a new sprite texture

The given filename is expected to end on _xx., e.g. asteroid_64x64x24.png.

fn draw(self: &Self,
layer: &Layer,
frame_id: u32,
position: Point2,
color: Color)
-> &Self

Draws a sprite onto the given layer.

fn draw_transformed(self: &Self,
layer: &Layer,
frame_id: u32,
position: Point2,
color: Color,
rotation: f32,
scale: Vec2)
-> &Self

Draws a sprite onto the given layer and applies given color, rotation and scaling.

fn width(self: &Self) -> f32

Returns the width of the sprite.

fn height(self: &Self) -> f32

Returns the height of the sprite.

fn num_frames(self: &Self) -> u32

Returns the number of frames of the sprite.