Struct radiant_rs::utils::AVec

pub struct AVec<T> {
    // some fields omitted
}

Vector supporting multiple non-blocking writers and a single blocking reader.

Methods

impl<T> AVec<T> where T: Default

fn new(capacity: u32) -> AVec<T>

Creates a new instance with given maximum capacity

fn push(&self, value: T) -> u32

Adds an element to the vector and returns the insert position. This function blocks get() and clear() until it returns.

fn map<'a>(&'a self, size: u32) -> AVecMapGuard<'a, T>

Maps a slice of the vector for r/w access. This function blocks get() and clear() until the mapped result goes out of scope.

fn len(&self) -> usize

Returns the current length of the vector. This function blocks get() and clear() until it returns.

fn clear(&self)

Clear the vector. This function blocks until it returns.

fn get<'a>(&'a self) -> AVecReadGuard<'a, T>

Maps the entire vector for read access. This function blocks until the mapped result goes out of scope.